requirements.txt
fastapi[standard]
uvicorn[standard]
langchain
langchain-core
langchain-google-genai
langchain-chroma
langchain-community
google-cloud-secret-manager
google-cloud-logging
google-auth
google-cloud-storage
firebase-admin
chroma
pypdf
fpdf
youtube-transcript-api
pytube
python-dotenv 
python-docx 
python-pptx
openpyxl 
pytube 